@@ -18,7 +18,8 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 0.0.1+1
 
 environment:
-  sdk: ">=2.16.0-134.1.beta <3.0.0"
+  sdk: ">=3.1.0"
+  flutter: ^3.10.0
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -29,60 +30,119 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
-  flutter_localizations: 
-    sdk: flutter
-  intl: ^0.17.0       
-  path_provider: ^2.0.8
-  formz: ^0.4.1
-  flutter_chat_ui:
-    git:
-      url: https://github.com/effektio/flyer-flutter_chat_ui.git
+
+  # UI
+  flutter_adaptive_scaffold: ^0.1.0
+  atlas_icons: ^0.0.12
+  # stories_editor:
+  #   path: "../../flutter_stories_editor"
+    # git:
+    #   url: https://github.com/acterglobal/flutter_stories_editor.git
+  flutter_chat_ui: ^1.6.9
   flutter_chat_types: ^3.3.2
-  google_fonts: ^2.3.1
-  flutter_svg: ^1.0.3
-  duration: ^3.0.11
-  image_picker: ^0.8.5
-  file_picker: ^4.5.1
-  open_file: ^3.2.1
-  cached_network_image: ^3.2.1
+  flutter_staggered_grid_view: ^0.6.2
+  flutter_svg: ">=1.0.0"
+  google_fonts: ^5.1.0
   flutter_link_previewer: ^3.0.1
   bubble: ^1.2.1
-  colorize_text_avatar: ^1.0.2
-  html_editor_enhanced: ^2.5.0
-  themed: ^4.0.0
-  effektio_flutter_sdk:
-    path: packages/rust_sdk
-  mime: ^1.0.2
-  flutter_colorpicker: ^1.0.3
-  emoji_picker_flutter: ^1.3.0
-  filesystem_picker: ^2.0.1
-  permission_handler: ^9.2.0
-  transparent_image: ^2.0.0
-  flutter_slidable: ^1.3.0
-  expandable_text: ^2.3.0
-  get: ^4.6.5
-  characters: ^1.2.0
-  sprintf: ^6.0.0
-  photo_view: ^0.14.0
-  flutter_parsed_text: ^2.2.1
-  avatar_stack: ^1.1.2
-  flutter_lorem: ^2.0.0
-  get_time_ago: ^1.1.5
-  implicitly_animated_reorderable_list: ^0.4.2
-  flutter_mentions: ^2.0.1
-  flutter_html: ^3.0.0-alpha.5
-  flutter_icons_null_safety: ^1.1.0
-  connectivity_plus: ^3.0.2
+  swipe_to: ^1.0.2
+  implicitly_animated_reorderable_list_2: ^0.5.0
+  image_picker: ^1.0.0
+  cross_file_image: ^1.0.0
   overlay_support: ^2.1.0
+  flutter_keyboard_visibility: ^5.4.0
+  expandable_text: ^2.3.0
+  cached_network_image: ^3.2.1
+  flutter_mentions:
+    git:
+      url: https://github.com/gtalha07/flutter_mentions.git
+  flutter_matrix_html:
+    git:
+      url: https://github.com/acterglobal/flutter_matrix_html.git
+  flutter_slidable: ^3.0.0
+  timelines: ^0.1.0
+  photo_view: ^0.14.0
+  acter_avatar:
+    # path: "../../acter-avatar"
+    git:
+      url: https://github.com/acterglobal/acter-avatar
+
+  # Code Generation
+  flutter_localizations:
+    sdk: flutter
+  intl: ^0.18.0
+  freezed: ^2.3.2
+  acter_flutter_sdk:
+    path: packages/rust_sdk
+
+  # State Management
+  get: ^4.6.5
+  flutter_riverpod: ^2.2.0
+  riverpod: ^2.3.6
+  # Routing
+  go_router: ^10.0.0
+
+  # widgets
+  table_calendar:
+    git:
+      url: https://github.com/acterglobal/table_calendar.git
+
+  settings_ui: '2.0.2'
+
+  # Utilities
+  path_provider: ^2.0.8
+  path: ^1.8.3
+  url_launcher: ^6.1.10
+  file_picker: ^5.2.5
+  emoji_picker_flutter: ^1.6.1
+  connectivity_plus: ^4.0.1
+  mime: ^1.0.2
+  permission_handler: ^10.2.0
+  sprintf: ^7.0.0
+  freezed_annotation: ^2.4.1
+  http_parser: ^4.0.2
+  flutter_custom_selector: ^0.0.3
+  jiffy: ^6.2.1
+  screenshot: ^2.1.0
+  http: ^1.0.0
+  string_validator: ^1.0.0
+  shake: ^2.2.0
+  crypto: ^3.0.2
+  flutter_math_fork: ^0.7.1
+  shared_preferences: ^2.2.0
+  image_picker_for_web: ^2.1.11
+  flutter_onboarding_slider: ^1.0.9
+  riverpod_infinite_scroll: ^1.0.9
+  infinite_scroll_pagination: ^3.2.0
+  awesome_notifications:
+    # path: "../../awesome_notifications"
+    git:
+      url: https://github.com/acterglobal/awesome_notifications
+      ref: acter-main
+  logging: ^1.2.0
+  args: ^2.4.2
+  breadcrumbs: ^0.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
   integration_test:
     sdk: flutter
-  icons_launcher: ^2.0.5
-  flutter_lints: ^1.0.4
-
+  icons_launcher: ^2.0.7
+  flutter_lints: ^2.0.2
+  build_runner: ^2.3.3
+  mockito: ^5.4.2
+  # The current build runner version is breaking when generating due to the issue 
+  # https://github.com/dart-lang/build/issues/3519, the most probable fix is to upgrade
+  # but that requires updated flutter SDK, temporarily we can add build_runner_core
+  # as a fix and be compatible with flutter 3.7 versions.
+  build_runner_core: ^7.2.7+1
+  flutter_gherkin:
+    # path: "../../flutter_gherkin"
+    git:
+      url: https://github.com/acterglobal/flutter_gherkin
+      ref: main
+  gherkin: ^3.1.0
 
 icons_launcher:
   image_path: "assets/icon/logo.png"
