@@ -1,46 +1,41 @@
-use crate::config::{LoginConfig, ENV_ROOM};
 use anyhow::Result;
-use clap::Parser;
-use effektio_core::ruma;
+use clap::Subcommand;
 
-mod fetch_news;
+mod execute;
+mod history;
+mod list;
+mod manage;
 mod mock;
-mod post_news;
 
-pub use fetch_news::FetchNews;
-pub use mock::Mock;
-pub use post_news::PostNews;
+pub use execute::ExecuteOpts;
+pub use history::HistoryOpts;
+pub use list::List;
+pub use manage::Manage;
+pub use mock::MockOpts;
 
-#[derive(clap::Subcommand, Debug)]
+#[derive(Subcommand, Debug)]
 pub enum Action {
-    /// Post News to a room
-    PostNews(PostNews),
-    /// Fetch News of the use
-    FetchNews(FetchNews),
-    /// Mock Data on fresh server
-    Mock(Mock),
+    /// List rooms
+    List(List),
     /// Room Management
     Manage(Manage),
+    /// Reviewing the room history
+    History(HistoryOpts),
+    /// Mock Data on fresh server
+    Mock(MockOpts),
+    /// Template Execution
+    Execute(ExecuteOpts),
 }
 
 impl Action {
     pub async fn run(&self) -> Result<()> {
         match self {
-            Action::PostNews(news) => news.run().await?,
-            Action::FetchNews(config) => config.run().await?,
+            Action::Manage(config) => config.run().await?,
             Action::Mock(config) => config.run().await?,
-            _ => unimplemented!(),
+            Action::List(config) => config.run().await?,
+            Action::History(config) => config.run().await?,
+            Action::Execute(config) => config.run().await?,
         };
         Ok(())
     }
-}
-
-/// Posting a news item to a given room
-#[derive(Parser, Debug)]
-pub struct Manage {
-    /// The room you want to post the news to
-    #[clap(short, long, parse(try_from_str), env = ENV_ROOM)]
-    pub room: Box<ruma::RoomId>,
-    #[clap(flatten)]
-    pub login: LoginConfig,
 }
