@@ -3,7 +3,15 @@ use serde_json::{json, value::to_raw_value};
 
 pub static PURPOSE_FIELD: &str = "m.room.purpose";
 pub static PURPOSE_FIELD_DEV: &str = "org.matrix.msc3088.room.purpose";
-pub static PURPOSE_TEAM_VALUE: &str = "org.effektio.team";
+pub static PURPOSE_TEAM_VALUE: &str = "global.acter.team";
+
+#[allow(non_snake_case)]
+pub mod KEYS {
+    pub static TASKS: &str = "tasks";
+    pub static CALENDAR: &str = "calendar";
+    pub static NEWS: &str = "news";
+    pub static PINS: &str = "pins";
+}
 
 const HISTORY: &str = r#"{
     "type": "m.room.history_visibility",
@@ -23,28 +31,27 @@ const ENCRYPTION: &str = r#"{
     }
 }"#;
 
-/// Generate the default set ot initial states for effektio teams
-pub fn default_effektio_group_states() -> Vec<Raw<AnyInitialStateEvent>> {
-    let mut v: Vec<Raw<AnyInitialStateEvent>> = [HISTORY, ENCRYPTION]
+/// Generate the default set ot initial states for acter teams
+pub fn default_acter_space_states() -> Vec<Raw<AnyInitialStateEvent>> {
+    let mut v = [HISTORY]
         .into_iter()
         .map(|a| serde_json::from_str::<Raw<AnyInitialStateEvent>>(a).expect("static don't fail"))
-        .collect();
+        .collect::<Vec<Raw<AnyInitialStateEvent>>>();
+    let r = to_raw_value(&json!({
+        "type": PURPOSE_FIELD_DEV,
+        "state_key": PURPOSE_TEAM_VALUE,
+        "content": {
+            "m.enabled": true,
+            "m.importance_level": 50
+        }
+    }))
+    .expect("static parsing of subtype doesn't fail");
 
-    v.push(Raw::from_json(
-        to_raw_value(&json!({
-            "type": PURPOSE_FIELD_DEV,
-            "state_key": PURPOSE_TEAM_VALUE,
-            "content": {
-                "m.enabled": true,
-                "m.importance_level": 50
-            }
-        }))
-        .expect("static parsing of subtype doesn't fail"),
-    ));
+    v.push(Raw::from_json(r));
     v
 }
 
-pub fn default_effektio_conversation_states() -> Vec<Raw<AnyInitialStateEvent>> {
+pub fn default_acter_convo_states() -> Vec<Raw<AnyInitialStateEvent>> {
     [HISTORY, ENCRYPTION]
         .into_iter()
         .map(|a| serde_json::from_str::<Raw<AnyInitialStateEvent>>(a).expect("static don't fail"))
@@ -55,15 +62,15 @@ pub fn initial_state_for_alias(
     main_alias: &OwnedRoomAliasId,
     alt_aliases: &Vec<OwnedRoomAliasId>,
 ) -> Raw<AnyInitialStateEvent> {
-    Raw::from_json(
-        to_raw_value(&json!({
-            "type": "m.room.canonical_alias",
-            "state_key": "",
-            "content": {
-                "alias": Some(main_alias),
-                "alt_aliases": alt_aliases,
-            }
-        }))
-        .expect("static doesn't fail"),
-    )
+    let r = to_raw_value(&json!({
+        "type": "m.room.canonical_alias",
+        "state_key": "",
+        "content": {
+            "alias": Some(main_alias),
+            "alt_aliases": alt_aliases,
+        }
+    }))
+    .expect("static doesn't fail");
+
+    Raw::from_json(r)
 }
