@@ -1,8 +1,8 @@
-use matrix_sdk::Session;
+use matrix_sdk::ruma::{OwnedDeviceId, OwnedUserId};
 use serde::{Deserialize, Serialize};
 use url::Url;
 
-/// Extensive Restore Token for Effektio Sessions
+/// Extensive Restore Token for Acter Sessions
 #[derive(Serialize, Deserialize)]
 pub struct RestoreToken {
     /// Was this registered per guest-account?
@@ -10,5 +10,15 @@ pub struct RestoreToken {
     /// Server homebase url
     pub homeurl: Url,
     /// Session to hand to client
-    pub session: Session,
+    pub session: CustomAuthSession,
+}
+
+#[derive(Serialize, Deserialize)]
+pub struct CustomAuthSession {
+    /// user id for login
+    pub user_id: OwnedUserId,
+    /// device id for login
+    pub device_id: OwnedDeviceId,
+    /// access token for login
+    pub access_token: String,
 }
